feat: revamp ui

+++ src/components/Profile/Profile.tsx
@@ -6,19 +6,19 @@ import {
   DropdownMenuItem,
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
-import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 import { LogOut, User } from "lucide-react";
 import { useRouter } from "next/navigation";
 import { signOut } from "next-auth/react";
 import { ROUTES } from "@/lib/constants/routes";
 import useUserData from "@/lib/hooks/useUserData";
+import Typography from "../Typography/Typography";
 
 export default function ProfileDropdown() {
   const router = useRouter();
-  const {nameInitials} = useUserData()
-
+  const { nameInitials } = useUserData();
 
   const handleLogout = () => {
+    if (typeof window === "undefined") return null;
     signOut({ callbackUrl: ROUTES.HOME });
     localStorage.clear();
   };
@@ -26,13 +26,24 @@ export default function ProfileDropdown() {
   return (
     <DropdownMenu>
       <DropdownMenuTrigger asChild>
-        <Avatar className="cursor-pointer bg-secondary-orange">
+        <div className="flex justify-center items-center max-md:w-10 max-md:h-10 rounded-full p-2 border border-[#452D1C] bg-[linear-gradient(to_bottom,#30251E_70%,#6F4D34_100%,#AD744A_100%)]">
+          <Typography
+            variant="body"
+            className="text-left opacity-100 font-bold text-[#C7C4C6]"
+          >
+            {nameInitials || "UN"}
+          </Typography>
+        </div>
+        {/* <Avatar className="cursor-pointer bg-secondary-orange">
           <AvatarImage src="/avatar.png" alt="User avatar" />
           <AvatarFallback className="bg-secondary-orange p-2 text-white">{nameInitials}</AvatarFallback>
-        </Avatar>
+        </Avatar> */}
       </DropdownMenuTrigger>
 
-      <DropdownMenuContent  align="end" className="w-44 bg-background-primary border-primary-orange">
+      <DropdownMenuContent
+        align="end"
+        className="w-44 bg-background-primary border-primary-orange"
+      >
         <DropdownMenuItem
           onClick={() => router.push(ROUTES.PROFILE)}
           className="cursor-pointer"
